# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Product Visualization Pipeline API for pet products
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  dimensionsL?: number | null;
  /** @nullable */
  dimensionsW?: number | null;
  /** @nullable */
  dimensionsH?: number | null;
  imageUrls: string[];
  /** @nullable */
  sourceUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductBody {
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  dimensionsL?: number | null;
  /** @nullable */
  dimensionsW?: number | null;
  /** @nullable */
  dimensionsH?: number | null;
  imageUrls?: string[];
  /** @nullable */
  sourceUrl?: string | null;
}

export interface UpdateProductBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  dimensionsL?: number | null;
  /** @nullable */
  dimensionsW?: number | null;
  /** @nullable */
  dimensionsH?: number | null;
  imageUrls?: string[];
  /** @nullable */
  sourceUrl?: string | null;
}

export interface AddImageBody {
  imageUrl: string;
}

export type RenderJobStatus =
  (typeof RenderJobStatus)[keyof typeof RenderJobStatus];

export const RenderJobStatus = {
  pending: "pending",
  claimed: "claimed",
  generating_3d: "generating_3d",
  compositing: "compositing",
  rendering_video: "rendering_video",
  completed: "completed",
  failed: "failed",
} as const;

export type RenderJobSceneType =
  (typeof RenderJobSceneType)[keyof typeof RenderJobSceneType];

export const RenderJobSceneType = {
  living_room: "living_room",
  bedroom: "bedroom",
  balcony: "balcony",
  garden: "garden",
  kitchen: "kitchen",
} as const;

export type RenderJobAnimalType =
  (typeof RenderJobAnimalType)[keyof typeof RenderJobAnimalType];

export const RenderJobAnimalType = {
  cat: "cat",
  dog: "dog",
  none: "none",
} as const;

export interface RenderJob {
  id: number;
  productId: number;
  productName: string;
  status: RenderJobStatus;
  stage?: string;
  sceneType: RenderJobSceneType;
  animalType: RenderJobAnimalType;
  priority: number;
  /** @nullable */
  workerId?: number | null;
  /** @nullable */
  modelUrl?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  progressPct?: number | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type CreateJobBodySceneType =
  (typeof CreateJobBodySceneType)[keyof typeof CreateJobBodySceneType];

export const CreateJobBodySceneType = {
  living_room: "living_room",
  bedroom: "bedroom",
  balcony: "balcony",
  garden: "garden",
  kitchen: "kitchen",
} as const;

export type CreateJobBodyAnimalType =
  (typeof CreateJobBodyAnimalType)[keyof typeof CreateJobBodyAnimalType];

export const CreateJobBodyAnimalType = {
  cat: "cat",
  dog: "dog",
  none: "none",
} as const;

export interface CreateJobBody {
  productId: number;
  sceneType: CreateJobBodySceneType;
  animalType: CreateJobBodyAnimalType;
  priority?: number;
}

export type UpdateJobStatusBodyStatus =
  (typeof UpdateJobStatusBodyStatus)[keyof typeof UpdateJobStatusBodyStatus];

export const UpdateJobStatusBodyStatus = {
  pending: "pending",
  claimed: "claimed",
  generating_3d: "generating_3d",
  compositing: "compositing",
  rendering_video: "rendering_video",
  completed: "completed",
  failed: "failed",
} as const;

export interface UpdateJobStatusBody {
  status: UpdateJobStatusBodyStatus;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  workerId?: number | null;
  /** @nullable */
  modelUrl?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  progressPct?: number | null;
  /** @nullable */
  completedAt?: string | null;
}

export interface NextJobResponse {
  job: RenderJob | null;
}

export type WorkerStatus = (typeof WorkerStatus)[keyof typeof WorkerStatus];

export const WorkerStatus = {
  online: "online",
  offline: "offline",
  busy: "busy",
} as const;

export interface Worker {
  id: number;
  name: string;
  /** @nullable */
  gpuModel?: string | null;
  status: WorkerStatus;
  /** @nullable */
  lastHeartbeat?: string | null;
  jobsCompleted: number;
  createdAt: string;
}

export interface RegisterWorkerBody {
  name: string;
  /** @nullable */
  gpuModel?: string | null;
}

export interface DashboardStats {
  totalProducts: number;
  totalJobs: number;
  completedJobs: number;
  pendingJobs: number;
  activeWorkers: number;
  successRate: number;
}

export interface JobStatusCount {
  status: string;
  count: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  label: string;
  /** @nullable */
  detail?: string | null;
  timestamp: string;
}

export type ListJobsParams = {
  status?: string;
  /**
   * @nullable
   */
  productId?: number | null;
};